--- test/App/BlogPostsSmokeTest.php
@@ -1,35 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace AppTest\App;
-
-use AppTest\WebTestCase;
-use Generator;
-use Psr\Http\Message\ResponseInterface;
-
-class BlogPostsSmokeTest extends WebTestCase
-{
-    /**
-     * @group        functional
-     * @dataProvider urlProvider
-     */
-    public function testPageIsSuccessful($url) : void
-    {
-        $response = $this->handleRequest('GET', $url);
-
-        self::assertInstanceOf(ResponseInterface::class, $response);
-        self::assertEquals(200, $response->getStatusCode());
-    }
-
-    public function urlProvider() : Generator
-    {
-        // Get all blog posts
-        foreach (new \DirectoryIterator('data/posts') as $file) {
-            if ($file->isFile()) {
-                // Return file name without extension
-                yield ['/blog/' . $file->getBasename('.md')];
-            }
-        }
-    }
-}
